move an internal methods to report slice
